+++ client/src/game/maps/collisions.ts
